+++ src/main/java/de/kreth/invoice/persistence/UserAdressRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import de.kreth.invoice.data.User;
 import de.kreth.invoice.data.UserAdress;
 
-public interface UserAdressRepository extends CrudRepository<UserAdress, Integer> {
+public interface UserAdressRepository extends CrudRepository<UserAdress, Long> {
 
     UserAdress findByUser(User user);
 }
