+++ src/main/java/de/kreth/invoice/persistence/UserBankRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import de.kreth.invoice.data.User;
 import de.kreth.invoice.data.UserBank;
 
-public interface UserBankRepository extends CrudRepository<UserBank, Integer> {
+public interface UserBankRepository extends CrudRepository<UserBank, Long> {
 
     UserBank findByUser(User user);
 }
